Improve keyboard(s)
I've removed the native Android theme and replaced it with something
flatter that feels less obtrusive.  Also redesigned the layouts to
give better visual distintion between keys like on a real keyboard.
Make the three different keyboard layouts easier to switch between
by keeping the keyboard switcher buttons the same size between
screens & highlighting the relevant button when on that screen, as
an indicator of where you are.  It's also slightly shorter.